common/requirements.txt
pymongo[srv]==4.2.0
kafka-python==2.0.2
